requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy